Fix a 1-in-a-million bug with rounding.
There are certain tasks which must be done periodically.  The set of such tasks
is maintained in a binary heap for efficiency, so that tasks may be executed in
order according to their execution time.  Rarely the periodic task timer would
fall on *exactly* the interval, at which point the "round_up" function would
round the interval to itself, entering a loop where the same task would run
endlessly.  This would effectively disable the daemon (and the fault-tolerance
code would remove it from the cluster), but it would need a manual restart.

This commit fixes the bug to ensure that the code always rounds to an interval
higher than the current interval.