Fix more clang static analysis problems.
Biggest change in this commit is a move from vfork/execve to
posix_spawn.  This moves the fd cleanup into the replicant process and
not the object process, and also eliminates what clang believed to be a
race denial of service vector.  FWIW, vfork claims that execve is safe
to call after vfork (and the only safe-to-call system call).